Rename pil/fonts_converter.py
