+++ scripts/pil/convert_fonts_8x8.py
@@ -6,7 +6,7 @@ ROOT_PATH = os.path.join(os.path.dirname(os.path.abspath(__file__)), '..', '..')
 ROOT_PATH = os.path.abspath(ROOT_PATH)
 sys.path.append(os.path.abspath(ROOT_PATH))
 
-from scripts.pil.dot_char import DotCharConverter
+from scripts.pil.fonts_converter import DotCharConverter
 
 
 if __name__ == '__main__':
